Update app.min.js
control4: Keypads: Change qty to a text field for name. Add the name to that device's field 12. Remove gang field. Remove keys field. Add a programming notes -> field 12. Add programming notes to Audio -> Field 12 on the amp. Add programming notes to video streams -> Field 12 on the controllers Lighting: Move Del Bank to the right side; to avoid misclicks Change position to be a static text. This will not change, but will be assigned as gangs get added and deleted. It will reduce the footprint of the gang field as well. Cameras: Remove the footage dropdown from the cameras. Add a name field and a note field. Change the storage estimate functionality. In realtime - analyze the amount of total cameras and how many NVRs will be needed. Use that data to determine the total storage amount. From that amount estimate how much each camera will be retained (in days) given a 5% motion triggered recording measurement. Security Change the dropdown to have generic names, D/W, Motion, Glass, Carbon, Smoke. Add a toggle on the top to choose between ADC and Resideo. IF ADC is selected, then use the current models (Manufacturer on the export is Qolsys for all sensors except the smoke, that one is DSC) If Resideo is selected, use the models: QC-PROSIXMINI, QC-PROSIXPIRV, QC-PROSIXGB, QC-PROSIXCOV, and QC-PROSIXCMBOV with Resideo as the Manufacturer. For the panels, use the current models for ADC and use QC-PROWLTOUCHC for the primary and remotes; Resideo is the Manufacturer. Reorder the panel dropdown so that primary is first. Change the Networking tab to Structured Wiring. Like the lighting tab, move the del bank button to the right to avoid accidental clicks. Remove the POE field on the tab. Remove the Locks Tab Remove the Misc Tab